gate define middleware design and role wise authorization

+++ app/Http/Controllers/backend/PermissionController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\PermissionStoreRequest;
 
 class PermissionController extends Controller
@@ -17,8 +18,9 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        $permissions= Permission::with(['module:id,module_name,module_slug'])->select('id','module_id','permission_name','permission_slug','updated_at')->latest('id')->paginate();
-        return view('admin.pages.permission.index',compact('permissions'));
+        Gate::authorize('index-permission'); //authorize this user to access/give assess to admin dashboard
+        $permissions = Permission::with(['module:id,module_name,module_slug'])->select('id', 'module_id', 'permission_name', 'permission_slug', 'updated_at')->latest('id')->paginate();
+        return view('admin.pages.permission.index', compact('permissions'));
     }
 
     /**
@@ -26,8 +28,9 @@ class PermissionController extends Controller
      */
     public function create()
     {
-        $modules=Module::select('id','module_name')->get();
-        return view('admin.pages.permission.create',compact('modules'));
+        Gate::authorize('create-permission'); //authorize this user to access/give assess to admin dashboard
+        $modules = Module::select('id', 'module_name')->get();
+        return view('admin.pages.permission.create', compact('modules'));
     }
 
     /**
@@ -35,10 +38,11 @@ class PermissionController extends Controller
      */
     public function store(PermissionStoreRequest $request)
     {
+        Gate::authorize('create-permission'); //authorize this user to access/give assess to admin dashboard
         Permission::updateOrCreate([
-            'module_id'=>$request->module_id,
-            'permission_name'=>$request->permission_name,
-            'permission_slug'=>Str::slug($request->permission_name)
+            'module_id' => $request->module_id,
+            'permission_name' => $request->permission_name,
+            'permission_slug' => Str::slug($request->permission_name)
         ]);
         Toastr::success('Permission Created Successfully');
         return redirect()->route('permission.index');
@@ -57,10 +61,11 @@ class PermissionController extends Controller
      */
     public function edit(string $id)
     {
-        $modules=Module::select('id','module_name')->get();
-        $permission=Permission::find($id);
+        Gate::authorize('edit-permission');//authorize this user to access/give assess to admin dashboard
+        $modules = Module::select('id', 'module_name')->get();
+        $permission = Permission::find($id);
 
-        return view('admin.pages.permission.edit',compact('modules','permission'));
+        return view('admin.pages.permission.edit', compact('modules', 'permission'));
     }
 
     /**
@@ -68,11 +73,12 @@ class PermissionController extends Controller
      */
     public function update(PermissionStoreRequest $request, string $id)
     {
-        $permission=Permission::find($id);
+        Gate::authorize('edit-permission');//authorize this user to access/give assess to admin dashboard
+        $permission = Permission::find($id);
         $permission->update([
-            'module_id'=>$request->module_id,
-            'permission_name'=>$request->permission_name,
-            'permission_slug'=>Str::slug($request->permission_name)
+            'module_id' => $request->module_id,
+            'permission_name' => $request->permission_name,
+            'permission_slug' => Str::slug($request->permission_name)
         ]);
         Toastr::success('Permission Updated Successfully');
         return redirect()->route('permission.index');
@@ -83,7 +89,8 @@ class PermissionController extends Controller
      */
     public function destroy(string $id)
     {
-        $permission=Permission::find($id);
+        Gate::authorize('delete-permission');//authorize this user to access/give assess to admin dashboard
+        $permission = Permission::find($id);
         $permission->delete();
         Toastr::success('Permission Delete Successfully');
         return redirect()->route('permission.index');
